# apps/server/openapi.yaml
openapi: 3.0.3
info:
  title: BillManager API
  description: |
    REST API for BillManager - a multi-tenant bill tracking application.

    ## Authentication

    API v2 uses JWT (JSON Web Token) authentication:
    1. Call `/api/v2/auth/login` with username/password to get tokens
    2. Include `Authorization: Bearer <access_token>` header in requests
    3. Include `X-Database: <database_name>` header to select workspace
    4. Use `/api/v2/auth/refresh` to get new access tokens when expired

    Access tokens expire after 15 minutes. Refresh tokens expire after 7 days.

  version: 4.3.2
  license:
    name: O'Saasy License
    url: https://osaasy.dev/
  contact:
    name: BillManager
    url: https://github.com/brdweb/billmanager
servers:
  - url: /api/v2
    description: API v2 (JWT Authentication)
...
paths:
...
  /bills:
    get:
      tags: [Bills]
      summary: List all bills
      operationId: listBills
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDatabase'
        - name: include_archived
          in: query
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: List of bills
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Bill'
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Bills]
      summary: Create a new bill
      operationId: createBill
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XDatabase'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BillCreateInput'
      responses:
        '201':
          description: Bill created
          headers:
            Idempotency-Replayed:
              $ref: '#/components/headers/IdempotencyReplayed'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                      message:
                        type: string
                      last_updated:
                        type: string
                        format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /users:
    get:
      tags: [Administration]
      summary: List users visible to the current administrator
      operationId: listAdminUsers
      x-capability: administration
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Users in the deployment or SaaS account
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ListDataResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Administration]
      summary: Create a user
      operationId: createAdminUser
      x-capability: administration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminUserCreateInput'}
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ObjectDataResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /databases:
    get:
      tags: [Administration]
      summary: List managed bill groups and their access members
      operationId: listAdminDatabases
      x-capability: administration
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Managed bill groups
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ListDataResponse'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
    post:
      tags: [Administration]
      summary: Create a bill group
      operationId: createAdminDatabase
      x-capability: administration
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AdminDatabaseCreateInput'}
      responses:
        '201':
          description: Bill group created
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ObjectDataResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Browser session accepted by endpoints using the shared auth decorator
  headers:
    IdempotencyReplayed:
      description: Present and true only when a stored mutation response was replayed
      schema:
        type: boolean
  parameters:
    XDatabase:
      name: X-Database
      in: header
      required: true
      description: |
        Database/workspace name to operate on. Use `_all_` only on endpoints
        that explicitly support cross-workspace access. Mutation idempotency is
        isolated by authenticated user and the affected database.
      schema:
        type: string
        example: personal
...
  responses:
    BadRequest:
      description: Invalid request or required workspace header missing
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Missing, invalid, or expired access token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated user cannot access the requested resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict:
      description: Resource changed or the mutation UUID was reused
      content:
        application/json:
          schema:
            oneOf:
              - $ref: '#/components/schemas/ConflictResponse'
              - $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string
          nullable: true
          description: Stable machine-readable code on new contract-aware errors
        details:
          type: object
          nullable: true
          additionalProperties: true
    ConflictResponse:
      type: object
      required: [success, error, code, conflict]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: string
        code:
          type: string
          enum: [resource_conflict]
        conflict:
          $ref: '#/components/schemas/ResourceConflict'
    ResourceConflict:
      type: object
      required: [entity, entity_id, reason, server_updated_at, server]
      properties:
        entity:
          type: string
          enum: [bill, payment, share]
        entity_id:
          type: integer
        reason:
          type: string
          enum: [modified]
        base_updated_at:
          type: string
          format: date-time
        server_updated_at:
          type: string
          format: date-time
        server:
          type: object
          additionalProperties: true
    MutationControl:
      type: object
      properties:
        client_mutation_id:
          type: string
          format: uuid
          description: |
            Optional idempotency UUID. Repeating the same user/database/request
            returns the original status and body with `Idempotency-Replayed: true`.
            A bill move remains scoped to the source database selected by the
            original request, so a retry is replayable after the bill moves.
        base_updated_at:
          type: string
          format: date-time
          description: >-
            Last server timestamp read by the client, encoded as RFC3339 with a
            UTC offset. Legacy clients may omit this field; when present it
            cannot be null, empty, or offset-free.
    ObjectDataResponse:
      type: object
      required: [success]
      additionalProperties: true
      properties:
        success: {type: boolean, enum: [true]}
        data:
          type: object
          additionalProperties: true
        message: {type: string}
    ListDataResponse:
      type: object
      required: [success, data]
      properties:
        success: {type: boolean, enum: [true]}
        data:
          type: array
          items:
            type: object
            additionalProperties: true
...
    Bill:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        amount:
          type: number
          nullable: true
        varies:
          type: boolean
        avg_amount:
          type: number
          description: Average payment amount (only for variable bills)
        frequency:
          type: string
          enum: [weekly, bi-weekly, biweekly, monthly, quarterly, yearly, custom]
        frequency_type:
          type: string
          enum: [simple, specific_dates, multiple_weekly]
        frequency_config:
          type: string
          description: JSON configuration for complex frequencies
        next_due:
          type: string
          format: date
        auto_payment:
          type: boolean
        icon:
          type: string
        type:
          type: string
          enum: [expense, deposit]
        account:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        reminder_enabled:
          type: boolean
        reminder_days:
          type: array
          items:
            type: integer
            minimum: 0
        database_id:
          type: integer
        last_updated:
          type: string
          format: date-time
        archived:
          type: boolean
        is_shared:
          type: boolean
          description: True if this bill is shared with the current user (recipient view)
        share_info:
          type: object
          nullable: true
          description: Share information (only present when is_shared is true)
          properties:
            share_id:
              type: integer
              description: ID of the share record
            owner_name:
              type: string
              description: Username of the bill owner
            my_portion:
              type: number
              nullable: true
              description: Amount the recipient owes
            my_portion_paid:
              type: boolean
              description: Whether recipient has marked their portion as paid
            my_portion_paid_date:
              type: string
              format: date-time
              nullable: true
              description: When recipient marked their portion as paid
    BillMutationFields:
      allOf:
        - $ref: '#/components/schemas/MutationControl'
        - type: object
          properties:
            database_id:
              type: integer
              description: Explicit destination workspace, required for create in `_all_`
            name:
              type: string
            amount:
              type: number
              nullable: true
            varies:
              type: boolean
              default: false
            frequency:
              type: string
              enum: [weekly, bi-weekly, biweekly, monthly, quarterly, yearly, custom]
              default: monthly
            frequency_type:
              type: string
              enum: [simple, specific_dates, multiple_weekly]
              default: simple
            frequency_config:
              type: string
              default: "{}"
            next_due:
              type: string
              format: date
            auto_payment:
              type: boolean
              default: false
            icon:
              type: string
              default: payment
            type:
              type: string
              enum: [expense, deposit]
              default: expense
            account:
              type: string
              nullable: true
            category:
              type: string
              nullable: true
            notes:
              type: string
              nullable: true
            reminder_enabled:
              type: boolean
              default: true
            reminder_days:
              type: array
              items:
                type: integer
                minimum: 0
    BillCreateInput:
      allOf:
        - $ref: '#/components/schemas/BillMutationFields'
        - type: object
          required: [name, next_due]
...
    AdminUserCreateInput:
      type: object
      required: [username, password]
      properties:
        username: {type: string}
        password: {type: string, format: password, minLength: 8}
        email: {type: string, format: email, nullable: true}
        role: {type: string, enum: [admin, user], default: user}
        database_ids:
          type: array
          items: {type: integer}
...
    AdminDatabaseCreateInput:
      type: object
      required: [name, display_name]
      properties:
        name: {type: string}
        display_name: {type: string}
        description: {type: string}
...